test(rust-autofixer): merge pinocchio visitor suites
